+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Rebroadcaster
 
+## Changed
+
+- Broadcast accepts an array of Observers
+
 ## [0.0.0] - 2023-07-21
 
 ### Added